gcc_mt_m11_v1/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]==2.5.3
python-dotenv==1.0.0
pytest==7.4.4

# Logging and utilities
structlog==24.1.0

# Pinecone Vector Database (primary service)
pinecone-client==3.0.0

# OpenAI Embeddings and LLM (secondary service)
openai==1.10.0

# Database (for production deployment with PostgreSQL)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Redis caching (optional)
redis==5.0.1

# JWT authentication
pyjwt==2.8.0

# Testing
pytest-asyncio==0.23.3
httpx==0.26.0

# Jupyter notebook support
jupyter==1.0.0
jupyterlab==4.0.10
ipykernel==6.29.0
matplotlib==3.8.2
